refactor: static files helper to middleware directory
