+++ examples/static-assets/index.php
@@ -3,11 +3,11 @@
 require __DIR__ . "/vendor/autoload.php";
 
 use Dumbo\Dumbo;
-use Dumbo\Helpers\StaticFiles;
+use Dumbo\Middleware\StaticFilesMiddleware;
 
 $app = new Dumbo();
 
-$app->use("/", StaticFiles::serve(__DIR__ . "/assets"));
+$app->use("/", StaticFilesMiddleware::serve(__DIR__ . "/assets"));
 
 // $app->get("/", StaticFiles::serve(__DIR__ . "/assets"));
